ecommerce-microservices/services/cart-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.3
PyJWT==2.8.0
httpx==0.26.0
